Custom font size closes #55
Possibilité de saisir la taille de la police au clavier (ComboBox éditables).

Support du status de fenêtre à taille maximisé dans l'enregistrement des dernières dimensions de MainScreen.

+++ src/main/java/fr/clementgre/pdf4teachers/utils/FontUtils.java
@@ -14,7 +14,7 @@ public class FontUtils {
             "Open Sans", "Jost", "Lato Black", "Lato", "Lato Light", "Roboto Medium", "Roboto", "Roboto Light", "Hind Guntur", "Shanti",
             "Karma", "Noto Serif", "Crimson Text", "Bitter",
             "Ubuntu Condensed", "Bellota", "Balsamiq Sans", "MuseoModerno",  "Averia Libre", "Indie Flower", "Sriracha", "Arrows");
-    public static final ObservableList<Integer> sizes = FXCollections.observableArrayList(6, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17, 18, 20, 22, 24, 26, 28, 30, 34, 38, 42, 46, 50);
+    public static final ObservableList<Double> sizes = FXCollections.observableArrayList(6d, 8d, 9d, 10d, 11d, 12d, 13d, 14d, 15d, 16d, 17d, 18d, 20d, 22d, 24d, 26d, 28d, 30d, 34d, 38d, 42d, 46d, 50d);
 
     public static Font getFont(String family, boolean italic, boolean bold, double size){
